+++ specs/lib/addPipelineExtrasSpec.js
@@ -17,28 +17,11 @@ describe('addPipelineExtras', () => {
                     type: WebGLConstants.VERTEX_SHADER,
                     uri: 'data:,'
                 }
-            },
-            images: {
-                sampleImage0: {
-                    extras: {
-                        compressedImage3DTiles: {
-                            s3tc: {
-                                uri: 'data:,'
-                            },
-                            etc1: {
-                                uri: 'data:,'
-                            }
-                        }
-                    }
-                }
             }
         };
         const gltfWithExtras = addPipelineExtras(gltf);
         expect(gltfWithExtras.buffers['sampleBuffer0'].extras._pipeline).toBeDefined();
         expect(gltfWithExtras.shaders['sample0VS'].extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images['sampleImage0'].extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images['sampleImage0'].extras.compressedImage3DTiles.s3tc.extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images['sampleImage0'].extras.compressedImage3DTiles.etc1.extras._pipeline).toBeDefined();
     });
 
     it('adds pipeline extras to glTF 2.0 assets', () => {
@@ -46,20 +29,6 @@ describe('addPipelineExtras', () => {
             buffers: [
                 {
                     byteLength: 100
-                }
-            ],
-            images: [
-                {
-                    extras: {
-                        compressedImage3DTiles: {
-                            s3tc: {
-                                uri: 'data:,'
-                            },
-                            etc1: {
-                                uri: 'data:,'
-                            }
-                        }
-                    }
                 }
             ],
             extensions: {
@@ -82,8 +51,5 @@ describe('addPipelineExtras', () => {
         const gltfWithExtras = addPipelineExtras(gltf);
         expect(gltfWithExtras.buffers[0].extras._pipeline).toBeDefined();
         expect(gltfWithExtras.extensions.KHR_techniques_webgl.shaders[0].extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images[0].extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images[0].extras.compressedImage3DTiles.s3tc.extras._pipeline).toBeDefined();
-        expect(gltfWithExtras.images[0].extras.compressedImage3DTiles.etc1.extras._pipeline).toBeDefined();
     });
 });