+++ CHANGES.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+### 3.0.0 - 2020-??-??
+
+* Breaking changes
+    * Dropped support for CRN and KTX textures. [#550](https://github.com/CesiumGS/gltf-pipeline/pull/550)
+    * Dropped support for `image.extras.compressedImage3DTiles`'s [#550](https://github.com/CesiumGS/gltf-pipeline/pull/550)
+* Added support for KTX2 textures. [#550](https://github.com/CesiumGS/gltf-pipeline/pull/550)
+
 ### 2.1.10 - 2020-05-11
 
 * Improved accessor min/max for Draco attributes. [#540](https://github.com/CesiumGS/gltf-pipeline/pull/540)